chore: kubernetes manifests generated from AuthProxyWorkload resource (#9)
These manifests are generated by the kubebuilder auth proxy workload resource. The controller-gen tool
that is part of kubebuilder reads the godoc annotations in api/v1alpha/authproxyworkload_types.go and writes
the kubernetes manifests. They are meant to be committed to the repository, even though they get regenerated.

This includes a change to the lint job to ensure that the make generate and make manifests are in sync with the
source repo.