test: create kubernetes api integration tests (#10)
Kubebuilder contains a tool called envtest which will set up a local kubernetes api server. These
test cases use that local api server to assert that the operator is correctly interacting with kubernetes
API. This allows us to test the controller for the AuthProxyWorkload resource as well as the mutating
webhooks for workloads.

The envtest tests are considered integration tests because they use an application running outside of
the testing framework, much like a database test. However they are not true end-to-end tests because
no kubernetes workload is ever scheduled or run.

This replaces the envtest test stubs generated by kubebuilder. Kubebuilder generates test that use
the Ginko framework. This project's owners prefer to use plain go testing tests as they are more accessible
to contributors.