feat: Run terraform scripts to provision infrastructure (#98)
Runs Terraform scripts to provision test infrastructure. This requires the 
developer have a valid Google Cloud project declared in an ignored file in the
project root: build.env. If the build.env does not exist when the developer runs 
make, then make will copy build.sample.env into place so the developer can fill in 
their information.

The terraform sources are located in the testinfra/ directory. The testinfra/run.sh 
script copies the terraform project directory into bin/tf/ so that terraform state files 
are in an ignored folder, out of danger from being committed to the repository.