feat: AuthProxyWorkload Reconcile no longer modifies workloads (#41 step 6) (#117)
This change removes all the old logic that modified the PodTemplateSpec of an existing workload.

Also it adjusts the state logic in the reconcile controller to so that it only reads, does not update
Workloads created by customers. The reconcile loop will only update the status on the AuthProxyWorkload
so that the status includes all listed deployments. 

And adds support for ReplicaSet workloads, so that our operator can better traverse the 
ownership hierarchy of a pod.