feat: Add support for Unix sockets. (#205)
The operator will configure the pod with a proxy that listens on a unix socket. The operator will mount the
parent directory of each instance's unixSocketPath to each container in the pod. When the proxy starts, it
will create a unix socket for the instance in that mounted directory.

Care should be take to ensure that the parent directory can safely be used across all containers in the pod.

Fixes #47