doc: Automatically generate API documentation for the AuthProxyWorkload (#167)
This will automatically generate API documentation for the AuthProxyWorkload CRD
and put it into the source code at docs/api.md. This uses the tool
crd-ref-docs. The api docs are generated
automatically during make generate so that the docs commited to the source code
are always in sync with go source.

Fixes #165