chore: Use package alias cloudsqlapi for ./internal/api/v1alpha (#255)
Update all files to use the import alias cloudsqlapi for the package
"github.com/GoogleCloudPlatform/cloud-sql-proxy-operator/internal/api/v1alpha1".

The API package name is going to change from v1alpha1 to v1 soon and we don't want a ton
of irrelevant changes. Thus, we want to use the package alias throughout the
code so that the package rename change is concise.