feat: Add support for Private Service Connect (#391)
Adds support to configure private service connect through the operator. This will set the proxy's "psc=true" parameter
to the instance connection string passed to the proxy.

In accordance with the validation in the proxy, Users may not set both PrivateIP and PSC at the same time. See root.go:517

Fixes #389