chore: Fix yaml formatter so that it only formats the config/ directory (#440)
We need to automatically format the yaml in config/ because some of it is generated
by controller-runtime and kubebuilder, some of it is hand-written. The formatter helps us
preserve the consistency and minimize noisy diffs when we upgrade the tools that generate yaml.

However, we don't want to format the hand-written yaml used elsewhere in this project as tool configuration
and example docs. The yaml formatter makes the yaml consistent, but not as readable.