chore: Fix the order of renovate package rules.
According to renovate documentation, the list should be ordered so that
low priority rules are at the beginning of the list and high priority rulesthe highest priority rule is at the end of the list. A dependency that matches
multiple rules will be added to the last group that it matches.