feat: Run the proxy container as Sidecar Init Container (#624)
Operator will configure proxy containers as sidecar containers in `PodSpec.InitContainers` when supported
by the kubernetes cluster. 

The operator checks the cluster version. If the cluster version is 1.29 or higher, the operator will configure the proxy
as an InitContainer sidecar container. If the cluster version is 1.28 or lower, the operator will configure proxy 
containers in the PodSpec.Container. 

This is especially helpful for Job and CronJob workloads. Sidecar containers are a new type of containers that start among the InitContainers, run through the lifecycle of the Pod and don’t block pod termination. Kubelet makes a best effort to keep them alive and running while other containers are running.

Fixes #381

Co-authored-by: Jonathan Hess <hessjc@google.com>
Co-authored-by: Jonathan Hess (he/him) <103529393+hessjcg@users.noreply.github.com>